+++ Sources/TagKit/Views/TagCapsule.swift
@@ -3,7 +3,7 @@
 //  TagKit
 //
 //  Created by Daniel Saidi on 2022-08-19.
-//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -11,49 +11,109 @@ import SwiftUI
 /**
  This view can be used to render tags within a capsule shape.
 
- Note that the view will not slugify the provided tag, since
- this would require it to use the same slug configuration as
- was used to generate the tag. Instead, just provide it with
- an already slugified tag.
+ The view will not slugify the provided tag string, only use
+ the content you provide it with.
  */
 public struct TagCapsule: View {
 
-    /**
-     Create a tag capsule.
-
-     - Parameters:
-       - tag: The tag to render.
-       - style: The style to use.
-     */
+    /// Create a tag capsule.
+    ///
+    /// - Parameters:
+    ///   - tag: The tag to render.
     public init(
-        tag: String,
-        style: TagCapsuleStyle
+        _ tag: String
     ) {
         self.tag = tag
-        self.style = style
     }
 
     private let tag: String
-    private let style: TagCapsuleStyle
+    
+    @Environment(\.tagCapsuleStyle)
+    private var style: TagCapsuleStyle
 
     public var body: some View {
         Text(tag)
             .padding(style.padding)
             .foregroundColor(style.foregroundColor)
-            .background(style.backgroundColor)
-            .clipShape(Capsule())
-            .overlay(Capsule().strokeBorder(
-                style.borderColor,
-                lineWidth: style.borderWidth))
+            .background(Capsule()
+                .strokeBorder(
+                    style.borderColor,
+                    lineWidth: style.borderWidth
+                )
+            )
+            .background(Capsule().fill(style.backgroundColor))
+            
     }
 }
 
-struct TagCapsule_Previews: PreviewProvider {
+#Preview {
 
-    static var previews: some View {
-        VStack {
-            TagCapsule(tag: "standard-tag", style: .standard)
-            TagCapsule(tag: "standard-selected-tag", style: .standardSelected)
-        }
+    VStack {
+        TagCapsule("standard-tag")
+        TagCapsule("standard-selected-tag")
+            .font(.body.bold())
+            .tagCapsuleStyle(.standardSelected)
+        TagCapsule("spider-man")
+            .tagCapsuleStyle(.init(
+                foregroundColor: .black,
+                backgroundColor: .red,
+                borderColor: .blue,
+                borderWidth: 4,
+                padding: .init(
+                    top: 10,
+                    leading: 20,
+                    bottom: 12,
+                    trailing: 20
+                )
+            ))
     }
+    .preferredColorScheme(.dark)
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+