+++ Sources/TagKit/Views/TagCapsuleStyle.swift
@@ -3,7 +3,7 @@
 //  TagKit
 //
 //  Created by Daniel Saidi on 2022-09-07.
-//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -16,7 +16,7 @@ import SwiftUI
 
  TagKit comes with a boring ``TagCapsuleStyle/standard`` and
  an even worse``TagCapsuleStyle/standardSelected`` style. Do
- make sure to define your own styles to make your tags pop!
+ design your own styles to make your tags pop!
  */
 public struct TagCapsuleStyle {
 
@@ -28,77 +28,101 @@ public struct TagCapsuleStyle {
        - backgroundColor: The foreground color to use, by default `.blue`.
        - borderColor: The border color to use, by default `.clear`.
        - borderWidth: The border width to use, by default `1`.
-       - padding: The padding to apply to the text.
+       - padding: The intrinsic padding to apply, by default a small padding.
      */
     public init(
         foregroundColor: Color = .primary,
         backgroundColor: Color = .blue,
         borderColor: Color = .clear,
         borderWidth: Double = 1,
-        padding: EdgeInsets = .init(top: 5, leading: 10, bottom: 7, trailing: 10)
+        padding: EdgeInsets? = nil
     ) {
+        var defaultPadding: EdgeInsets
+        #if os(tvOS)
+        defaultPadding = .init(top: 12, leading: 20, bottom: 14, trailing: 20)
+        #else
+        defaultPadding = .init(top: 5, leading: 10, bottom: 7, trailing: 10)
+        #endif
+        
         self.foregroundColor = foregroundColor
         self.backgroundColor = backgroundColor
         self.borderColor = borderColor
         self.borderWidth = borderWidth
-        self.padding = padding
+        self.padding = padding ?? defaultPadding
     }
 
-    /**
-     The foreground color to use, by default `.primary`.
-     */
+    /// The foreground color to use.
     public var foregroundColor: Color
 
-    /**
-     The foreground color to use, by default `.blue`.
-     */
+    /// The foreground color to use.
     public var backgroundColor: Color
 
-    /**
-     The border color to use, by default `.clear`.
-     */
+    /// The border color to use.
     public var borderColor: Color
 
-    /**
-     The border width to use, by default `1`.
-     */
+    /// The border width to use.
     public var borderWidth: Double
 
-    /**
-     The padding to apply to the text.
-     */
+    /// The padding to apply to the text.
     public var padding: EdgeInsets
 }
 
 public extension TagCapsuleStyle {
 
-    /**
-     Get the standard selected tag badge style.
-
-     You can override this style to change the default style.
-     */
+    /// A standard tag capsule style.
+    ///
+    /// You can set this style to change the global default.
     static var standard = TagCapsuleStyle(
-        foregroundColor: .white,
-        backgroundColor: .blue,
-        borderColor: .clear)
+        foregroundColor: .black,
+        backgroundColor: .gray,
+        borderColor: .clear
+    )
 
-    /**
-     Get the standard tag badge style.
-
-     You can override this style to change the default style.
-     */
+    /// A standard, selected tag capsule style.
+    ///
+    /// You can set this style to change the global default.
     static var standardSelected = TagCapsuleStyle(
         foregroundColor: .white,
-        backgroundColor: .green,
-        borderColor: .clear)
+        backgroundColor: .black,
+        borderColor: .white
+    )
 }
 
-struct TagCapsuleStyle_Previews: PreviewProvider {
+public extension View {
 
-    static var previews: some View {
-        VStack {
-            TagCapsule(tag: "standard-tag", style: .standard)
-            TagCapsule(tag: "standard-selected-tag", style: .standardSelected)
-        }
+    /// Apply a ``TagCapsule`` style to the view hierarchy.
+    func tagCapsuleStyle(
+        _ style: TagCapsuleStyle
+    ) -> some View {
+        self.environment(\.tagCapsuleStyle, style)
+    }
+}
+
+private extension TagCapsuleStyle {
+
+    struct Key: EnvironmentKey {
+
+        public static var defaultValue: TagCapsuleStyle = .standard
+    }
+}
+
+public extension EnvironmentValues {
+
+    /// This value can bind to a line spacing picker config.
+    var tagCapsuleStyle: TagCapsuleStyle {
+        get { self [TagCapsuleStyle.Key.self] }
+        set { self [TagCapsuleStyle.Key.self] = newValue }
+    }
+}
+
+
+#Preview {
+
+    VStack {
+        TagCapsule("standard-tag")
+        TagCapsule("standard-selected-tag")
+            .font(.body.bold())
+            .tagCapsuleStyle(.standardSelected)
+            .shadow(radius: 0, x: 0, y: 2)
     }
 }